Duplicate code clean up
