+++ BiomarinPKU/BiomarinPKU-Study/AppDelegate.swift
@@ -51,8 +51,8 @@ class AppDelegate: SBAAppDelegate, RSDTaskViewControllerDelegate {
                                               accent: RSDColorMatrix.shared.colorKey(for: .palette(.rose),
                                                                                      shade: .dark))
     static let designSystem = RSDDesignSystem(version: 1,
-                                              colorRules: PKUColorRules(palette: colorPalette, version: 1),
-                                              fontRules: PKUFontRules(version: 1))
+                                              colorRules: RSDColorRules(palette: colorPalette, version: 1),
+                                              fontRules: FontRules(version: 1))
     
     /**
      TODO: mdephillips 5/24/19
@@ -83,7 +83,7 @@ class AppDelegate: SBAAppDelegate, RSDTaskViewControllerDelegate {
         Localization.insert(bundle: mainBundle, at: 0)
         
         // Set up font rules.
-        RSDStudyConfiguration.shared.fontRules = PKUFontRules(version: 0)
+        RSDStudyConfiguration.shared.fontRules = FontRules(version: 0)
         
         // Setup reminders
         RSDStudyConfiguration.shared.shouldShowRemindMe = true
@@ -233,62 +233,9 @@ extension AppDelegate {
     }
     
     class func setupFooter(_ footer: RSDNavigationFooterView) {
-        guard let primaryButtonFont = (self.designSystem.fontRules as? PKUFontRules)?.font(for: .primary) else { return }
+        guard let primaryButtonFont = (self.designSystem.fontRules as? FontRules)?.font(for: .primary) else { return }
         
         footer.backButton?.titleLabel?.font = primaryButtonFont
         footer.nextButton?.titleLabel?.font = primaryButtonFont
-    }
-}
-
-open class PKUColorRules: RSDColorRules {
-    
-}
-
-open class PKUFontRules: RSDFontRules {
-    
-    public let latoRegularName      = "Lato-Regular"
-    public let latoBoldName         = "Lato-Bold"
-    public let latoBlackName        = "Lato-Black"
-    public let latoLightName        = "Lato-Light"
-    public let latoItalicName       = "Lato-Italic"
-    public let latoBoldItalicName   = "Lato-BoldItalic"
-    public let latoLightItalicName  = "Lato-LightItalic"
-    
-    override open func font(for buttonType: RSDDesignSystem.ButtonType) -> RSDFont {
-        switch buttonType {
-        case .primary:
-            return RSDFont(name: latoBoldName, size: 20)!
-        case .secondary:
-            return RSDFont(name: latoBoldName, size: 20)!
-        }
-    }
-    
-    /// Returns the font to use for a given text type.
-    ///
-    /// - parameter textType: The text type for the font.
-    /// - returns: The font to use for this text.
-    override open func font(for textType: RSDDesignSystem.TextType) -> RSDFont {
-        switch textType {
-        case .heading1:
-            return RSDFont(name: latoBoldName, size: 24)!
-        case .heading2:
-            return RSDFont(name: latoBoldName, size: 18)!
-        case .heading3:
-            return RSDFont(name: latoBoldName, size: 16)!
-        case .heading4:
-            return RSDFont(name: latoRegularName, size: 14)!
-        case .fieldHeader:
-            return RSDFont(name: latoBoldName, size: 16)!
-        case .body:
-             return RSDFont(name: latoRegularName, size: 18)!
-        case .bodyDetail:
-            return RSDFont(name: latoRegularName, size: 16)!
-        case .small:
-            return RSDFont(name: latoRegularName, size: 14)!
-        case .microHeader:
-            return RSDFont(name: latoBoldName, size: 12)!
-        case .counter:
-            return RSDFont(name: latoLightName, size: 80)!
-        }
     }
 }
